backend/requirements.txt
fastapi[standard]
langchain==0.3.4
langchain-community==0.3.3
langchain-core==0.3.25
langchain-ollama==0.2.0
loguru==0.7.2
streamlit==1.39.0
faiss-cpu==1.9.0
python-dotenv